wobbly: Explicitly stop Libanimation D-Bus service when disabled
Rather than just dropping our final reference to the service, explicitly
stop it (which causes it to unexport all its objects and unown its D-Bus
name). Previously, it did this when the server object was finalised, but
since `wobbly.js` is interpreted code, we can’t guarantee when the
garbage collector will be run, and hence finalisation may happen seconds
after the final reference being dropped in the JS. In addition, other
code may still hold references to the server object, preventing
finalisation.

This allows the retry timeout workaround to be removed.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T29887