codeView: Remove sessions on extension disable
When the app and the toolbox are opened, a new coding session exists to
map this relation. This coding session shows and hides the corresponding
window depending on the state and does the flip animation.

If we've a coding session and we disable the extension, we should remove
the toolbox window because this window is not useful if it's not paired
with the app itself.

This patch removes all coding sessions on disable and also sets the app
actor to the default state (not rotated and without colour effect).

https://phabricator.endlessm.com/T29887