hackfx: Add code_view_effect
This patch adds a new C library to the project, hackfx. This library
will have clutter effect that are used in the extension and right now
are in the gnome-shell endless branch.

This first commit copies the shell_code_view_effect class from the shell
here and it's renamed to hackfx_code_view_effect.

https://phabricator.endlessm.com/T30450