wobbly: Full javascript wobbly window effect
This patch replaces the custom EOSShellWobbly clutter effect with the
new WobblyEffect.

This new wobbly effect is full javascript so it doesn't depend on
specific gnome-shell changes and it should work on any modern GNOME
desktop.

The new wobblyFx.js file is based on this gnome-shell plugin:
https://github.com/hermes83/compiz-alike-windows-effect

It's a copy of the effects.js file with the commonUtils.js functions
copied at the top and the enable/disable copied from extension.js.

The Prefs function implement the settings that we've right now in the
eos-hack-extension and provide default values for settings that we don't
manage.

https://phabricator.endlessm.com/T30485