codeView: Link toolbox and app windows
This patch modifies the behavior for the alt-tab and the appIconBar to
handle the toolbox and app window like it was only one window.

The toolbox is not shown as part of the clubhouse and activating the app
will bring to top the toolbox window when the codeView session is
flipped.

This replaces the behavior implemented in the shell window tracker but
instead of doing that low level, it's done in the interface.

The performance will be worst, but the code session list is a small
list, one or two, so iterate over all every time we do the alt tab it's
not a big issue.

https://phabricator.endlessm.com/T30450