appDisplay: Launch clubhouse using DBus API
The toolbox is inside the katamari so if we've a toolbox opened and
click on the hack icon it just bring up the toolbox and do not open the
clubhouse.

This patch calls directly to the activate DBus method to ensure that the
clubhouse window is shown in any situation.

The running dot in the icon is also updated accordingly, to be shown
only when the clubhouse window is there.

https://phabricator.endlessm.com/T30450