codeView: Proxy AppIcon.app to handle CodingSessions
This patch adds a proxy for the app attribute of AppIcon so we can
modify the list of windows shown in the dash AppIcon menu.

It also provides a new activate method if the app is the clubhouse, to
avoid bringing to the front a toolbox window when clicking on the
clubhouse icon.

https://phabricator.endlessm.com/T30450