codeView: Show/Hide clubhouse icon on dash
This patch improves the behavior of the hack icon on the dash bar. The
hack icon will be visible only if the clubhouse window is opened.

We should override this because the hack-toolbox is part of the
clubhouse flatpak and gnome desktop shows the clubhouse icon by default
if there's a toolbox running. This is something that can happen if the
user opens the sidetrack and flip, without the clubhouse window.

https://phabricator.endlessm.com/T30924