utils: Reset hack desktop modifications
The clubhouse changed the background and cursor theme on hack mode
enable / disable change. The hack mode has been removed and the
clubhouse doesn't change the background and cursor theme anymore.

This patch adds some code to the extension to ensure to reset the
modified values from a previous clubhouse version and will avoid
problems in the future if we remove the background images from the
clubhouse.

https://phabricator.endlessm.com/T30736