clubhouse: Fallback for Flatpak lib gir
This commit adds a Fallback for when the gir1.2-flatpak-1.0 is not
installed. In that case we try to find the real clubhouse directory
using the `flatpak info` command line.

Maybe in the future we can get this information using the clubhouse API,
instead of trying to detect where it's installed.

https://phabricator.endlessm.com/T31029