clubhouse: Fix url parsing on markup
The 'a' tag is not valid for pango on the shell, so we should remove
every 'a' tag on the text before calling Pango.parse_markup to avoid an
Error.

This patch just removes all 'a' tags and leave the text inside the tag.
The clubhouse shell notification code parses all links and convert them
into clickable widgets, so this will fix the issue with links not
working.

https://phabricator.endlessm.com/T31032