Disable forward compatibility
The forward compatibility is enabled by default on major linux
distributions like Ubuntu or Fedora, this keys is true:

disable-extension-version-validation

To avoid breaking the system with the hack extension on update, this
patch introduces a check that will disable the extension if the version
is greater than expected.

We'll need to review and adapt the code for future versions before the
release to keep the extension working with OS updates.

https://phabricator.endlessm.com/T31135